Move cipher code into the new file ibmca_cipher.c
We moved all cipher code from e_ibmca.c into a new created file
ibmca_cipher.c. Also added this file to Makefile so it can be built.
