+++ src/Makefile.am
@@ -1,6 +1,6 @@
 lib_LTLIBRARIES=ibmca.la
 
-ibmca_la_SOURCES=e_ibmca.c e_ibmca_err.c
+ibmca_la_SOURCES=e_ibmca.c e_ibmca_err.c ibmca_cipher.c
 ibmca_la_LIBADD=-ldl
 ibmca_la_LDFLAGS=-module -version-info 0:2:0 -shared -no-undefined -avoid-version
